Grunt: remove svn:ignore property from `tests/phpunit/` for `phpunit.xml`.
This was an artifact of pre-2.1 versions of BuddyPress where the unit tests folder lived at the root
level of the project, and so we needed to support ignoring custom phpUnit build files.

The root `trunk` folder now has an ignore for `phpunit.xml`, and that's where custom phpUnit build
files should be placed going forwards.

This commit is part of a series that will re-organise BuddyPress trunk to use a Grunt build process. See #5160


git-svn-id: https://buddypress.svn.wordpress.org/trunk@8361 cdf35c40-ae34-48e0-9cc9-0c9da1808c22